# Conflicts:
#	FWLog.Data/FWLog.Data.csproj
#	FWLog.Data/Repository/GeneralCtx/NotaFiscalRepository.cs
#	FWLog.Data/UnitOfWork.cs
#	OracleScripts/Inserir Registros.sql

+++ FWLog.Data/Models/Lote.cs
@@ -7,9 +7,9 @@ namespace FWLog.Data.Models
     public class Lote
     {
         [Key]
-        public int IdLote { get; set; }
-        public int IdLoteStatus { get; set; }
-        public int IdNotaFiscal { get; set; }
+        public long IdLote { get; set; }
+        public long IdLoteStatus { get; set; }
+        public long IdNotaFiscal { get; set; }
         public DateTime DataCompra { get; set; }
         public DateTime DataRecebimento { get; set; }
         public int QuantidadePeca { get; set; }
