+++ FWLog.Data/Models/UnidadeMedida.cs
@@ -5,7 +5,7 @@ namespace FWLog.Data.Models
     public class UnidadeMedida
     {
         [Key]
-        public int IdUnidadeMedida { get; set; }
+        public long IdUnidadeMedida { get; set; }
         public string Descricao { get; set; }
     }
 }
